pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pw-workflow-runner"
version = "0.1.0"
description = "Workflow executor for the Parallel Works ACTIVATE platform"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Parallel Works" }
]
dependencies = [
    "parallelworks-client @ file:///home/mattshax/pw-sdk/python",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
pw-workflow-runner = "pw_workflow_runner.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pw_workflow_runner"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = false
